Rewrite swissprot parser
Not finished yet.
This will remove the dependency of BioPython's SwissProtParser,
which is not updated with the new SwissProt protocol.
We also do not need all the fields in the .dat file so we can
just parse the fields we need for annotation.

Doing this will make the package easier to maintain.